+++ lib/lotus/source/adapters/ecto/dialects/mysql.ex
@@ -7,8 +7,8 @@ defmodule Lotus.Source.Adapters.Ecto.Dialects.MySQL do
   alias __MODULE__.EditorConfig
   alias Lotus.Query.Statement
   alias Lotus.Source.Adapters.Ecto.Dialects.Default
-  alias Lotus.SQL.FilterInjector
-  alias Lotus.SQL.SortInjector
+  alias Lotus.Source.Adapters.Ecto.SQL.FilterInjector
+  alias Lotus.Source.Adapters.Ecto.SQL.SortInjector
 
   @default_statement_timeout_ms 5_000
 
@@ -499,7 +499,7 @@ defmodule Lotus.Source.Adapters.Ecto.Dialects.MySQL do
 
   @impl true
   def transform_statement(%Statement{text: sql} = statement) do
-    alias Lotus.SQL.Transformer
+    alias Lotus.Source.Adapters.Ecto.SQL.Transformer
 
     new_sql =
       sql