+++ test/lotus/source/adapters/ecto/sql/validator_test.exs
@@ -1,7 +1,7 @@
-defmodule Lotus.SQL.ValidatorTest do
+defmodule Lotus.Source.Adapters.Ecto.SQL.ValidatorTest do
   use Lotus.Case, async: false
 
-  alias Lotus.SQL.Validator
+  alias Lotus.Source.Adapters.Ecto.SQL.Validator
 
   describe "validate/2 with postgres" do
     test "accepts PostgreSQL-specific ILIKE syntax" do